requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# AI/ML libraries
openai>=1.3.0
google-generativeai>=0.3.0

# Image processing
Pillow>=10.0.0

# PDF generation
reportlab>=4.0.0

# Email functionality
# (uses built-in smtplib, no additional package needed)

# Utilities
python-dateutil>=2.8.2
